fixed db seed

+++ src/server/migrations/20190804002546-create-cities.js
@@ -20,11 +20,11 @@ module.exports = {
         allowNull: false,
       },
       longitude: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       latitude: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       population: {
